requirements.txt
python-telegram-bot==21.7
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
google-cloud-secret-manager==2.20.2
google-auth>=2.0.0,<3.0.0
python-json-logger==2.0.7
